fix(adaptable-components): Fixed component re-rendering issues
This moves props and ref into the react context shared by the adaptable component and the adaptations. This, and some React.memo, will have the generated component tree not being reconstructed when e.g. props change. Reconstricting that tree (since the components are _generated_) will lead to new components, hence React will reset the entire tree downstream and recreate it again. So this reduces that to practically never happen.

Signed-off-by: Gustaf Räntilä <g.rantila@gmail.com>